extract client and printer
