+++ pkg/kubectl/cmd/top.go
@@ -17,6 +17,7 @@ limitations under the License.
 package cmd
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"strings"
@@ -25,8 +26,10 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/kubernetes/pkg/kubectl"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
+	client "k8s.io/kubernetes/pkg/client/unversioned"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/unversioned"
+	"k8s.io/kubernetes/pkg/kubectl/metricsutil"
 )
 
 // TopOptions is the start of the data required to perform the operation. As new fields are added, add them here instead of
@@ -85,8 +88,8 @@ func NewCmdTop(f *cmdutil.Factory, out io.Writer) *cobra.Command {
 		ArgAliases: argAliases,
 	}
 	cmd.Flags().StringP("selector", "l", "", "Selector (label query) to filter on")
-	// TODO: add support for --all-namespaces
 	cmd.Flags().Bool("containers", false, "If present, print usage of containers within a pod. Ignored if the requested resource type is \"node\".")
+	// TODO: add support for --all-namespaces
 	cmdutil.AddInclude3rdPartyFlags(cmd)
 	return cmd
 }
@@ -108,10 +111,50 @@ func RunTop(f *cmdutil.Factory, out io.Writer, cmd *cobra.Command, args []string
 		return err
 	}
 
-	// TODO: refactor to builder?
-	err = kubectl.PrintMetrics(out, cli, cmdNamespace, false, printContainers, params, args)
+	err = PrintMetrics(out, cli, cmdNamespace, false, printContainers, params, args)
 	if err != nil {
 		return err
 	}
 	return nil
+}
+
+func PrintMetrics(out io.Writer, cli *client.Client, cmdNamespace string, allNamespaces bool,
+		  printContainers bool, params map[string]string, args []string) error {
+	resType, err := GetResourceKind(args[0])
+	if err != nil {
+		return err
+	}
+	name := ""
+	if len(args) > 1 {
+		name = args[1]
+	}
+	// TODO: check for a heapster config?
+	heapsterClient := metricsutil.DefaultHeapsterMetricsClient(cli)
+	printer := metricsutil.NewTopCmdPrinter(out)
+
+	switch resType {
+	case api.Kind("Node"):
+		metrics, err := heapsterClient.GetNodeMetrics(name, params)
+		if err != nil {
+			return err
+		}
+		return printer.PrintNodeMetrics(metrics)
+	case api.Kind("Pod"):
+		metrics, err := heapsterClient.GetPodMetrics(cmdNamespace, name, params)
+		if err != nil {
+			return err
+		}
+		return printer.PrintPodMetrics(metrics, printContainers)
+	}
+	return errors.New("resource not supported")
+}
+
+var GetResourceKind = func(resourceType string) (unversioned.GroupKind, error) {
+	switch resourceType {
+	case "node", "nodes":
+		return api.Kind("Node"), nil
+	case "pod", "pods":
+		return api.Kind("Pod"), nil
+	}
+	return unversioned.GroupKind{}, errors.New("unknown resource requested")
 }
